Move mem to backend
